% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\alias{htype}
\alias{htype<-}
\alias{inode}
\alias{inode<-}
\alias{ip}
\alias{ip<-}
\alias{nhtypes}
\alias{op}
\alias{op<-}
\title{Getters and setters}
\usage{
htype(f)

htype(f) <- value

nhtypes(f)

inode(f)

inode(f) <- value

ip(f)

ip(f) <- value

op(f)

op(f) <- value
}
\arguments{
\item{f}{any function}

\item{value}{character string}
}
\description{
\itemize{
  \item htype - Get or set types for unary function
  \item nhtypes - The number of types a function has (2 for unary functions)
  \item ip,op - get input and output types
  \item inode - get/set input nodes
}
}

