% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasard.R
\name{active_node}
\alias{active_node}
\title{Wrap a pure function in a monad ...}
\usage{
active_node(fpure, fival, foval, fstat, feff)
}
\arguments{
\item{fpure, fival, foval, fstat, feff}{some functions}
}
\value{
wrapped function
}
\description{
Wrap a pure function in a monad ...
}

