% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_class}
\alias{add_class}
\title{Prepend a class to the class vector}
\usage{
add_class(x, ...)
}
\arguments{
\item{x}{anything}

\item{...}{classes to add}
}
\value{
x with the new class prepended
}
\description{
Prepend a class to the class vector
}

