% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{are_composable}
\alias{are_composable}
\title{Check if two functions are composable}
\usage{
are_composable(f, g)
}
\arguments{
\item{f, g}{unary class functions}
}
\value{
logical
}
\description{
Inputs are considered to be composable if they are functions and are either
untyped or have composable types. Typed functions (f,g) are composable if 1)
f's output is not NA (i.e. is not a sink), 2) g's input is not NA (i.e. is
not a source), and 3) f's output equals g's input.
}

