% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom.R
\name{basic_functions}
\alias{basic_functions}
\alias{blank}
\alias{execute}
\alias{false}
\alias{id}
\alias{nocache}
\alias{nothing}
\alias{true}
\title{Basic functions}
\usage{
true(...)

false(...)

nothing(...)

blank()

id(x, ...)

execute(FUN, ...)

nocache(op, ...)
}
\arguments{
\item{...}{anything}

\item{x}{anything}

\item{FUN}{any function}

\item{op}{command for cache function}
}
\description{
\itemize{
  \item true - ignore all input and return TRUE
  \item false - ignore all input and return FALSE
  \item nothing - ignore all input and do nothing
  \item blank - takes no arguments and returns nothing
  \item id - returns first argument unchanged, ignores others
  \item execute - pass ... to FUN and return the results
  \item cache - a cache function that does nothing
}
}

