% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_functions}
\alias{cache_functions}
\alias{make_datcache}
\alias{make_memcache}
\alias{nocache}
\title{Cache functions}
\usage{
nocache(op, ...)

make_memcache()

make_datcache(filename)
}
\arguments{
\item{op}{command for cache function}

\item{...}{arguments sent to del, chk, put, and get within the cacher}

\item{filename}{filename for cachers that store data in files}
}
\description{
Cache functions
}

