% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{classcheck}
\alias{classcheck}
\title{Check the htype}
\usage{
classcheck(CLASS, ...)
}
\arguments{
\item{CLASS}{expected class}

\item{...}{any number of objects}
}
\value{
TRUE if all elements in ... have class CLASS
}
\description{
Check the htype
}

