% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{combine}
\alias{combine}
\title{Combine takes any number of arguments, but yields a function with a constant
number of arguments.}
\usage{
combine(...)
}
\arguments{
\item{...}{input functions from which types are induced}
}
\value{
combining function
}
\description{
Combine takes any number of arguments, but yields a function with a constant
number of arguments.
}

