% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{compose}
\alias{compose}
\title{Make composition of functions}
\usage{
compose(...)
}
\arguments{
\item{...}{two or more functions}
}
\value{
A function that is a composition of the input functions
}
\description{
Input functions should be ordered from innermost to outtermost. All
arguments are passed to the innermost function, whose formal arguments are
transferred exactly to the composed function.
}
\details{
The composed function is a member of the environment in which `compose` was
called. The composition does not copy the constituent functions, but rather
searches for them in the parent frame. So, given `foo <- compose(f,g,h)`
calling `foo()` is exactly equivalent to calling `h(g(f))()` in foo's
parental environment.
}
\examples{
compose(runif, mean, abs, log)
}

