% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{composition_test}
\alias{composition_test}
\title{Assert two functions can be composed}
\usage{
composition_test(f, g)
}
\arguments{
\item{f, g}{two functions}
}
\value{
logical
}
\description{
Assert two functions can be composed
}

