% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{connect}
\alias{connect}
\title{Connect nodes into a pipeline}
\usage{
connect(x)
}
\arguments{
\item{x}{string describing functional pathway}
}
\description{
Pathways are expressed using a simple language. Given nodes A, B, C, and D:
}
\details{
* 'A' - performs no operations
 * 'A --> B' := `h_inode(B) <- A` in the parent frame

Multiple inputs are space delimited

* 'A B --> C' := `h_inode(C) <- list(A, B)`

Any number of nodes can be chained

* 'A --> B --> C' := `h_inode(B) <- A; h_inode(C) <- B`

Parentheses can be used to denote branching pathways

* '(A --> B) (C --> D) E'
   1. `h_inode(B) <- A`
   2. `h_inode(D) <- C`
   3. `h_inode(E) <- list(B, D)`
}
\examples{
library(pied)
library(magrittr)
A <- hwell('a')
B <- hpipe('a -> b')
C <- hpipe('b -> c')
D <- hpipe('b -> c -> d')

h_fun(A) <- function() 'a'
h_fun(B) <- function(x) paste0(x, 'b')
h_fun(C) <- function(x) paste0(x, 'c')
h_fun(D) <- function(x,y) sprintf('(\%s)(\%s)d', x, y)

connect('(A --> B) (A --> B --> C) --> D')

D()
}

