% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{default_fun}
\alias{default_fun}
\title{Create a default (NULL returning) function for a given type}
\usage{
default_fun(type)
}
\arguments{
\item{type}{functional type}
}
\description{
Create a default (NULL returning) function for a given type
}
\seealso{
default_inode
}

