% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{default_inode}
\alias{default_inode}
\title{Create a default inode object for a given type}
\usage{
default_inode(type)
}
\arguments{
\item{type}{functional type}
}
\description{
Create a default inode object for a given type
}
\seealso{
default_fun
}

