% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{effify}
\alias{effify}
\title{Add a side-effect to the function call}
\usage{
effify(f, effect = nothing)
}
\arguments{
\item{f}{a unary function (a -> b)}

\item{effect}{a function of a and b}
}
\value{
a closure that calls effect(a,b) without affecting f(a)
}
\description{
Add a side-effect to the function call
}

