% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasard.R
\name{fcompose}
\alias{fcompose}
\title{Make composition of functions}
\usage{
fcompose(...)
}
\arguments{
\item{...}{two or more functions}
}
\value{
A function that is a composition of the input functions
}
\description{
Make composition of functions
}

