% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{flift}
\alias{flift}
\title{Generic flift function: (a -> b) -> (F a -> F b)}
\usage{
flift(f, s)
}
\arguments{
\item{f}{function of type (a -> b)}

\item{s}{function of type (F a -> b -> F b)}
}
\value{
a state_bound datum
}
\description{
Generic flift function: (a -> b) -> (F a -> F b)
}

