% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{flift.state_bound}
\alias{flift.state_bound}
\title{(a -> b) -> (F a -> F b)}
\usage{
\method{flift}{state_bound}(f, s)
}
\arguments{
\item{f}{function of type (a -> b)}

\item{s}{function of type (F a -> b -> F b)}
}
\value{
a state_bound datum
}
\description{
(a -> b) -> (F a -> F b)
}

