% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{hnode}
\alias{hnode}
\title{Build function}
\usage{
hnode(f, .itype = "*", .otype = "*", .inode = NULL, .onode = NULL,
  .val = true, .iwrap = id, .owrap = id, .pass = execute,
  .fail = nothing, .effect = nothing, .cacher = nocache, ...)
}
\arguments{
\item{f}{pure function}

\item{.itype, .otype}{input and output types}

\item{.inode, .onode}{input and output nodes}

\item{.val}{(a -> Bool) - determines in input is correct}

\item{.iwrap, .owrap}{input and output wrappers}

\item{.pass}{function called if val returns TRUE}

\item{.fail}{function called if val returns FALSE}

\item{.effect}{a function of a and b}

\item{.cacher}{the caching function}

\item{...}{constant arguments that will be written into the closure}
}
\description{
Build function
}

