% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_function}
\alias{is.effectual}
\alias{is.tuple}
\alias{is.typed}
\alias{is.unary}
\alias{is.validated}
\alias{is_function}
\title{is.* functions for checking classes}
\usage{
is.tuple(x)

is.unary(x)

is.typed(x)

is.validated(x)

is.effectual(x)
}
\arguments{
\item{x}{anything}
}
\value{
logical
}
\description{
is.* functions for checking classes
}

