% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{keep_split}
\alias{keep_split}
\title{Like splits on a pattern but keeps the pattern}
\usage{
keep_split(x, p, dummy = "@@")
}
\arguments{
\item{x}{a character vector}
}
\value{
p a pattern (can be a regular expression)
}
\description{
Like splits on a pattern but keeps the pattern
}
\examples{
keep_split(c('a,b', ',acat,,eh'), ',')
keep_split('(a){strangely(punctuated..*sentence', '[^\\\\w]')
}

