% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ltypecheck}
\alias{ltypecheck}
\title{Check the htype for a list}
\usage{
ltypecheck(x, type)
}
\arguments{
\item{x}{a list}

\item{type}{expected type}
}
\value{
logical
}
\description{
Check the htype for a list
}

