% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{make_cacher}
\alias{make_cacher}
\title{Create a caching function}
\usage{
make_cacher(fchk, fdel, fput, fget)
}
\arguments{
\item{fchk}{determines if a cached value exists}

\item{fdel}{deletes any cached value}

\item{fput}{caches the input}

\item{fget}{returns the cached value}
}
\value{
a caching function
}
\description{
Create a caching function
}

