% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_effector}
\alias{make_effector}
\title{Create a new effector function}
\usage{
make_effector(type, f = nothing)
}
\arguments{
\item{type}{the type of the function}

\item{f}{function that should be of the specified input type}
}
\value{
an effector function
}
\description{
Create a new effector function
}

