% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_validator}
\alias{make_validator}
\title{Create a new validator}
\usage{
make_validator(type, f = true)
}
\arguments{
\item{type}{the type of the function}

\item{f}{function that should be of the specified input type}
}
\value{
a validator function
}
\description{
Create a new validator
}
\seealso{
Other makers: \code{\link{make_cacher}},
  \code{\link{make_effector}}
}

