% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{memify}
\alias{memify}
\title{Get output for a function}
\usage{
memify(f, cacher = execute)
}
\arguments{
\item{f}{any function}

\item{cacher}{the caching function}
}
\value{
function ouptut
}
\description{
Get output for a function
}

