% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nhtypes}
\alias{nhtypes}
\title{The number of types a function has}
\usage{
nhtypes(f)
}
\arguments{
\item{f}{a function with the 'typed' class}
}
\value{
integer
}
\description{
For a unary function, this will be two: input and output types. Currently,
this is the only legal kind of typed function. But eventually I may change
this.
}

