% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{parsubset}
\alias{parsubset}
\title{Extract a subset from a tuple}
\usage{
parsubset(x, i)
}
\arguments{
\item{x}{a tuple (or list)}

\item{i}{indices to extract from x}
}
\value{
a subset of x
}
\description{
Extract a subset from a tuple
}

