% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.active_node}
\alias{print.active_node}
\title{Print active_node}
\usage{
\method{print}{active_node}(x, ...)
}
\arguments{
\item{x}{input function}

\item{...}{additional arguments, will be ignored}
}
\description{
Print active_node
}

