% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.typed}
\alias{print.unary}
\title{Hasard print functions}
\usage{
\method{print}{unary}(x, ...)

\method{print}{typed}(x, ...)
}
\arguments{
\item{x}{input function}

\item{...}{additional arguments, will be ignored}
}
\description{
Hasard print functions
}

