% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.unary}
\alias{print.unary}
\title{Print unary function}
\usage{
\method{print}{unary}(x, ...)
}
\arguments{
\item{x}{input function}

\item{...}{additional arguments, will be ignored}
}
\description{
Print unary function
}

