% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.wired_node}
\alias{print.wired_node}
\title{Print wired_node}
\usage{
\method{print}{wired_node}(x, ...)
}
\arguments{
\item{x}{input function}

\item{...}{additional arguments, will be ignored}
}
\description{
Print wired_node
}

