% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rclass}
\alias{rclass}
\title{A function that recusively determines class from an object}
\usage{
rclass(.)
}
\arguments{
\item{.}{any thing}
}
\value{
character a class definition
}
\description{
A function that recusively determines class from an object
}

