% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{runall}
\alias{runall}
\title{Run functions for their side effects}
\usage{
runall(x, ...)
}
\arguments{
\item{x}{a function or a list of functions}

\item{...}{arguments sent to x}
}
\description{
Run functions for their side effects
}

