% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasard.R
\name{tee}
\alias{tee}
\title{Split output of a function into downstream functions}
\usage{
tee(f, g, h)
}
\arguments{
\item{f, g, h}{input functions}
}
\value{
a list of two functions
}
\description{
Split output of a function into downstream functions
}

