% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{tuple}
\alias{tuple}
\title{Build a tuple (based on a list) from arguments}
\usage{
tuple(...)
}
\arguments{
\item{...}{data to include in the output tuple}
}
\value{
a list with 'tuple' class
}
\description{
Build a tuple (based on a list) from arguments
}

