% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{tuplesubset}
\alias{tuplesubset}
\title{Build a function for subsetting a tuple}
\usage{
tuplesubset(x, i)
}
\arguments{
\item{x}{a tuple (or list)}

\item{i}{indices to extract from x}
}
\value{
a function for subsetting a tuple
}
\description{
Build a function for subsetting a tuple
}

