% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{tuplify}
\alias{tuplify}
\title{Build a tuple (based on a list) from arguments}
\usage{
tuplify(...)
}
\arguments{
\item{...}{data to include in the output tuple}
}
\value{
a list with 'tuple' class
}
\description{
Build a tuple (based on a list) from arguments
}

