% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{typify}
\alias{typify}
\title{Set the type of a unary function}
\usage{
typify(f, itype = "a", otype = "b")
}
\arguments{
\item{f}{input function}

\item{itype, otype}{input and output types}
}
\value{
a typed f
}
\description{
Set the type of a unary function
}

