% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unify}
\alias{unify}
\title{Collapse all arguments into unary closure}
\usage{
unify(f, ...)
}
\arguments{
\item{f}{arbitrary function}

\item{...}{additional arguments}
}
\description{
Collapse all arguments into unary closure
}

