% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{valify}
\alias{valify}
\title{Add input type checking to a typed function}
\usage{
valify(f, val = true, pass = execute, fail = nothing)
}
\arguments{
\item{f}{a typed, unary function}

\item{val}{(a -> Bool) - determines in input is correct}

\item{pass}{function called if val returns TRUE}

\item{fail}{function called if val returns FALSE}
}
\value{
f
}
\description{
Add input type checking to a typed function
}

