% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vlift}
\alias{vlift}
\title{Lift a value into a context}
\usage{
vlift(value, state = list(ok = TRUE))
}
\arguments{
\item{value}{anything}

\item{state}{anything}
}
\value{
the value inside a context
}
\description{
Lift a value into a context
}

