% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{wired_node}
\alias{wired_node}
\title{Contextualize an active function}
\usage{
wired_node(f, inode = NULL, onode = NULL)
}
\arguments{
\item{f}{input active_node}

\item{inode, onode}{input and output nodes}
}
\value{
a wired_node
}
\description{
Contextualize an active function
}

