% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{are_composable}
\alias{are_composable}
\title{Check if two functions are composable}
\usage{
are_composable(f, g)
}
\arguments{
\item{f, g}{unary class functions}
}
\value{
logical
}
\description{
Check if two functions are composable
}

