% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom.R
\name{basic_functions}
\alias{basic_functions}
\alias{execute}
\alias{false}
\alias{nothing}
\alias{true}
\title{Basic functions}
\usage{
true(...)

false(...)

nothing(...)

execute(FUN, ...)
}
\arguments{
\item{...}{anything}

\item{FUN}{any function}
}
\description{
\itemize{
  \item true - ignore all input and return TRUE
  \item false - ignore all input and return FALSE
  \item nothing - ignore all input and do nothing
  \item execute - pass ... to FUN and return the results
}
}

