% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom.R
\name{basic_functions}
\alias{basic_functions}
\alias{blank}
\alias{execute}
\alias{fail}
\alias{false}
\alias{id}
\alias{nothing}
\alias{true}
\alias{warn}
\title{Basic functions}
\usage{
true(...)

false(...)

nothing(...)

fail(...)

warn(...)

blank()

id(x, ...)

execute(FUN, ..., envir = parent.frame())
}
\arguments{
\item{...}{anything}

\item{x}{anything}

\item{FUN}{any function}

\item{envir}{environment in which to execute}
}
\description{
\itemize{
  \item true - ignore all input and return TRUE
  \item false - ignore all input and return FALSE
  \item nothing - ignore all input and do nothing
  \item fail - throws an exception for any input
  \item warn - calls a warning for any input
  \item blank - takes no arguments and returns nothing
  \item id - returns first argument unchanged, ignores others
  \item execute - pass ... to FUN and return the results
}
}

