% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_functions}
\alias{cache_functions}
\alias{make_datcache}
\alias{memcache}
\alias{nocache}
\title{Cache functions}
\usage{
nocache(id = NA)

memcache(id = NA)

make_datcache(directory)
}
\arguments{
\item{id}{unique id for the cache, if it will be the filename}

\item{directory}{directory where caches files are kept}
}
\description{
Cache functions
}

