% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{make_validator}
\alias{make_validator}
\title{Create a new validator}
\usage{
make_validator(f, vclass)
}
\arguments{
\item{f}{function with type: f :: vclass -> logical}

\item{vclass}{the input class}
}
\value{
a unary, validator function
}
\description{
Create a new validator
}

