% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{hnode}
\alias{node}
\title{Build function}
\usage{
hnode(f, ...)
}
\arguments{
\item{f}{pure function}

\item{...}{see Arguments}
}
\description{
Build function
}
\section{Arguments}{


\itemize{
  \item .itype input type
  \item .otype output type
  \item .inode input nodes
  \item .val (a -> Bool) - determines in input is correct
  \item .pass function called if val returns TRUE
  \item .fail function called if val returns FALSE
  \item .effect a function of a and b
  \item .cacher the caching function
}
}

