% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{h_args}
\alias{h_args<-}
\alias{h_cacher}
\alias{h_cacher<-}
\alias{h_delete}
\alias{h_delete<-}
\alias{h_effect}
\alias{h_effect<-}
\alias{h_fail}
\alias{h_fail<-}
\alias{h_fun}
\alias{h_fun<-}
\alias{h_inode}
\alias{h_inode<-}
\alias{h_pass}
\alias{h_pass<-}
\alias{h_val}
\alias{h_val<-}
\alias{hnode}
\alias{node}
\title{Build function}
\usage{
hnode(type, ...)

h_fun(h)

h_inode(h)

h_val(h)

h_pass(h)

h_fail(h)

h_effect(h)

h_cacher(h)

h_args(h)

h_delete(h)

h_fun(h) <- value

h_inode(h) <- value

h_val(h) <- value

h_pass(h) <- value

h_fail(h) <- value

h_effect(h) <- value

h_cacher(h) <- value

h_args(h) <- value

h_delete(h) <- value
}
\arguments{
\item{type}{a functional type ala Haskell}

\item{...}{see Arguments}

\item{h}{hnode function}

\item{value}{right hand value for assignment}
}
\description{
Build function
}
\section{Arguments}{


\itemize{
  \item type function type
  \item f function
  \item inode input nodes
  \item val (a -> Bool) - determines in input is correct
  \item pass function called if val returns TRUE
  \item fail function called if val returns FALSE
  \item effect a function of a and b
  \item cacher the caching function
}
}

