% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{rclass}
\alias{rclass}
\title{A function that recusively determines type from an object}
\usage{
rclass(x)
}
\arguments{
\item{x}{any thing}
}
\value{
character a class definition
}
\description{
A function that recusively determines type from an object
}

