% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{runall}
\alias{runall}
\title{Run functions for their side effects}
\usage{
runall(f, ..., envir = parent.frame())
}
\arguments{
\item{f}{a function or a list of functions}

\item{...}{arguments sent to f}

\item{envir}{environment in which to execute f}
}
\description{
Run functions for their side effects
}

