% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasard.R
\name{unify}
\alias{unify}
\title{Reduce a function to a single input, single output function}
\usage{
unify(f, itype = "*", otype = "*", ...)
}
\arguments{
\item{f}{input function}

\item{itype, otype}{input and output types}

\item{...}{constant arguments that will be written into the closure}
}
\value{
a unary class function
}
\description{
Reduce a function to a single input, single output function
}

